+++ src/main/java/com/ghgande/j2mod/modbus/io/FastByteArrayInputStream.java
@@ -42,23 +42,23 @@ public class FastByteArrayInputStream extends InputStream {
     /**
      * Actual position pointer into the input buffer.
      */
-    protected int pos;
+    int pos;
 
     /**
      * Marked position pointer into the input buffer.
      */
-    protected int mark;
+    int mark;
     /**
      * Input buffer <tt>byte[]</tt>.
      */
-    protected byte[] buf;
+    byte[] buf;
 
     /**
      * Creates an input stream.
      *
      * @param buffer the data to read.
      */
-    public FastByteArrayInputStream(byte[] buffer) {
+    FastByteArrayInputStream(byte[] buffer) {
         buf = buffer;
         count = buffer.length;
         pos = 0;
@@ -82,7 +82,7 @@ public class FastByteArrayInputStream extends InputStream {
 
     public int read() throws IOException {
         logger.debug("read()");
-        logger.debug("count={}", count, pos);
+        logger.debug("count={} pos={}", count, pos);
         return (pos < count) ? (buf[pos++] & 0xff) : (-1);
     }
 
@@ -126,7 +126,7 @@ public class FastByteArrayInputStream extends InputStream {
     public void mark(int readlimit) {
         logger.debug("mark()");
         mark = pos;
-        logger.debug("mark={}", mark, pos);
+        logger.debug("mark={} pos={}", mark, pos);
     }
 
     public void reset() {
@@ -176,4 +176,4 @@ public class FastByteArrayInputStream extends InputStream {
         return count;
     }
 
-}// class FastByteArrayInputStream
+}