+++ src/main/java/com/ghgande/j2mod/modbus/io/ModbusTransaction.java
@@ -20,6 +20,8 @@ import com.ghgande.j2mod.modbus.ModbusException;
 import com.ghgande.j2mod.modbus.msg.ModbusRequest;
 import com.ghgande.j2mod.modbus.msg.ModbusResponse;
 
+import java.util.Random;
+
 /**
  * Interface defining a ModbusTransaction.
  * <p>
@@ -38,6 +40,7 @@ public abstract class ModbusTransaction {
     protected ModbusResponse response;
     boolean validityCheck = Modbus.DEFAULT_VALIDITYCHECK;
     int retries = Modbus.DEFAULT_RETRIES;
+    private Random random = new Random(System.nanoTime());
     static int transactionID = Modbus.DEFAULT_TRANSACTION_ID;
 
     /**
@@ -123,7 +126,7 @@ public abstract class ModbusTransaction {
      * getTransactionID -- get the next transaction ID to use.
      * @return next transaction ID to use
      */
-    public int getTransactionID() {
+    synchronized public int getTransactionID() {
         /*
          * Ensure that the transaction ID is in the valid range between
          * 1 and MAX_TRANSACTION_ID (65534).  If not, the value will be forced
@@ -136,6 +139,16 @@ public abstract class ModbusTransaction {
             transactionID = 1;
         }
         return transactionID;
+    }
+
+    /**
+     * A useful method for getting a random sleep time based on an increment of the retry count and retry sleep time
+     *
+     * @param count Retry count
+     * @return Random sleep time in milliseconds
+     */
+    long getRandomSleepTime(int count) {
+        return (Modbus.RETRY_SLEEP_TIME / 2) + (long) (random.nextDouble() * Modbus.RETRY_SLEEP_TIME * count);
     }
 
     /**