+++ RELEASE_NOTES.md
@@ -107,3 +107,7 @@
 * Propagate ModbusTCPListener timeout to the TCPSlaveConnection #59 - thanks javi
 * Not possible to override handleRequest in ModbusSlave #57 - thanks javi
 * Extending ModbusSlave exceptions list
+
+## Version 2.3.8
+* Added a sleep timer to the TCP/UDP/serial transaction retry loops
+* Added the timeout value to the serial